+++ CHANGELOG.md
@@ -1,6 +1,21 @@
-### Version 2.30 [In Development]
+### Version 2.31 [In Development]
+**All WoW Versions:**  
+ElvUI 13.31+ compatibility  
 **Retail:**  
+Defaults for new Queue Status (Position, Font, Size)  
+Defaults for new addon compartment (Position, Size, Strata)  
+**Wrath Classic:**  
+Defaults for new Queue Status (Position, Font, Size)  
+**Classic:**  
+Fixed installer error in Layout  
+___
+### Version 2.30 [02-05-2023]
+**All WoW Versions:**  
+ElvUI 13.30+ compatibility  
+**Retail:**  
+Added initial setup for Addon Compartment font and position  
 Updated Shadow&Light profile  
+Updated TOC for 10.1  
 ___
 ### Version 2.29 [19-04-2023]
 **All WoW Versions:**